Migrated data to files, renamed test and added yaf and error.
